书籍评分更新，个人评分汇总，并添加网络评分

+++ java web type/src/main/java/com/jkb/entity/Booked.java
@@ -9,7 +9,7 @@ public class Booked {
 
     private long Sno;
 
-    private Integer code;
+    private long code;
 
     private byte buy;
